Adder/Subtractor test sequences
In files named with "raw", there are both decimal inputs in double and posit input in 32-bit with 2-bit Exponent.
Files named with "Posit", they are the input files which are simulated with. The number in the middle of the filename indicates how many sets of test sequences have been used.
Files named with "Rxxx" is the result of multiplication.